Remove CLion dependencies from `debugger` module
Now, `debugger` module contains code related to native debugger only.
Previously, it used some CLion only code. Now all such code moved into `clion` module.
To make sure that `debugger` module doesn't depend on CLion, it is compiled with [nativeDebug](https://plugins.jetbrains.com/plugin/12775-native-debugging-support) plugin dependency instead of CLion when base IDE is IDEA.

Also, duplicate code from `DebugRunner` and `DebugRunnerLegacy` was moved into a single place

Part of #3453
