+++ gradle-192.properties
@@ -1,6 +1,7 @@
 ideaVersion=IU-2019.2.3
 clionVersion=CL-2019.2.3
 
+nativeDebugPluginVersion=192.5728.99
 graziePluginVersion=2019.2-6.2.stable
 psiViewerPluginVersion=192-SNAPSHOT
