+++ gradle-193.properties
@@ -1,6 +1,7 @@
 ideaVersion=IU-2019.3
 clionVersion=CL-2019.3
 
+nativeDebugPluginVersion=193.5233.57
 graziePluginVersion=2019.3-6.2.stable
 psiViewerPluginVersion=193-SNAPSHOT
